Associate hash-table per listener manager
I think this is a more flexible and functional interface. osc-listen
returns a osc-listener-manager, which is associated with a server and
has its own table of patterns and callbacks.

Now when registering a listener you need to give it a manager. It adds
the necessary information to the table and returns the manager. The
relationship to table of callbacks and listener thread is a lot clearer
now.